+ rc: colors: switch to 16 color mode, brighter and better
256 color mode is just too dark, we tried overriding colors at least for
greyscale, this works but we would have to do it for all colors because
the entire span is dark, we'd have to shift everything, working this out
sounds like a pain.  for now setting 16 colors takes care of most of the
problem.  tried 88 colors but they aren't documented anywhere and have
to figure them out by rgb with trial and error, forget that (rgb.txt
documents only some colors that work, most of those do not work, more
work with 256 color of course but not even all of those work)