+ rc: force-set http referrer to a nonsense one to avoid leakage
hopefully this won't break anything

using value 2 is "use the fake one" and we changed it not to have a
controversial word, but the prior value was "use the correct one" which
always leaks the previous url (including any credentials inside the
arguments)

this may break something, we'll see.  the other option is to have it
always set to the same page as is being loaded.  not sure what that
might break but it's the next thing to try if this doesn't work