+ rc: bind: have w and W toggle wide or narrow margins
even with 9 the screen is often too wide.  some things just look bad
especially at smaller font sizes.  will be good to have this available
for alternate more narrow rendering of the page, with a value that isn't
too excessive.  would be nice if we could add and remove to the value
with a keypress...