- mailcap: image: pqiv: run in its own session so unaffected by tstp
if we don't do this, for some reason we cannot run it independently,
which we want to do, sometimes we will suspend the parent and look up
another page.  seems to be little downside because even if we don't ever
switch away from it, we'll just exit it and it will go back to the last
window, whether it was in a new session or not

would be nice to figure out why this happens though... looked at the
code and don't see a reason.  is it because it's using blocked=1
automatically? this is normally flagged for needsterminal probably, but
is it implied? need to read more code, but no time and this will work
fine for the moment