+ build: rc: libevent: enable now that it works for suspend
upstream seems to have fixed suspend for libevent in b53bcce, so we will
try it out and see if it feels any faster.  there are a couple places
where elinks seems to block ui while doing something else related to
networking; in theory those places can be reduced or vanish with
libevent

not using libev because it just doesn't seem to be actively developed.
it's maintained perhaps, but by just one or two people and not actively
developed.  tmux, which is used heavily for similar purposes, is using
libevent, and there are lots of commits all the time.  libev guy says
libevent is junk, but we'll see how it works out in practice.