+ urlviewrc: spawn elinks directly, as sensible-browser uses nohup
urlview by default uses /etc/urlview/url_handler.sh, which calls elinks
as "elinks nohup" which makes elinks go into dump mode.  probably
because it opens /dev/null and copies it to fd0 which makes elinks think
it is not a tty

the reason it uses nohup is because DISPLAY is set.  we could work
around it with a wrapper that unsets it, but it seems better to use just
elinks directly via rcfile, which always works and can fork off a
gui-browser if we want

it's the only reason we need an rcfile for it though.  that's ok since
it gets installed along with the installrc of elinks, its primary user